Update app.py
Added some headers that work logically such that if both filters are applied, the global cannot be entered. But if a single home team is selected to analyze, it filters the data to show all opponents vs a single team.

When therer are both a home team and away team specified, it also shows the score of the game and only shows the local dataframes for both in that single game.